Moved method from script helpers
